requirements.txt
fastapi
pydantic
uvicorn
openai-whisper
python-multipart